pet_insurance_rag/config/requirements.txt
# 펫보험 RAG 시스템 - 필수 패키지 목록
# 설치: pip install -r config/requirements.txt

# === AI/ML 패키지 ===
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# === 벡터 데이터베이스 ===
faiss-cpu>=1.7.4
numpy>=1.24.0

# === 데이터 처리 ===
pandas>=2.0.0
tiktoken>=0.5.0

# === 웹 인터페이스 ===
gradio>=4.0.0

# === 유틸리티 ===
python-dotenv>=1.0.0
requests>=2.31.0

# === 개발 도구 (선택사항) ===
jupyter>=1.0.0
notebook>=6.5.0